move config to separate file
